Introduce GridPlot as a return type of plot()
We introduce a class that abstracts a grid of subplots. It should
provide more object-oriented methods to manipulate matplotlib grid
subplots (mostly after creation).

The plot() method of Hypothesis, Experiment now returns an instance
of GridPlot rather than a ndarray of matplotlib AxesSubplot.

The implementation is still incomplete as we need to add more
functionalities, but it works almost the same as the previous basic
behavior of plot().