Add dummy __next__ methods to RunList and Hypothesis
When a pandas DataFrame or Series object contains a `RunList` or
`Hypothesis` object (with dtype=object), its pretty-print formatter
will convert them into a sequence of `Run` objects. We would rather want
to see the string representation (`__repr__`) in the pandas table.

Since pandas's `pprint_thing()` checks if the object has the attribute
`__next__` before checking if it is a sequence, we can use this as
a simple hack to avoid such an unwanted formatting.