Use multiprocess library in place of multiprocessing
multiprocess is a third-party fork of python standard library
multiprocessing (as a drop-in replacement). This library handles
non-picklable objects better, which is a very common trouble
when it comes to fork-based multiprocessing, and works better with
ipython/jupyter.

In light of common use cases, I think there is no reason not to
make multiprocess.pool (not multiprocessing.pool) as the sensible
default choice for the parallelism backend.