Provide presets for ex.plot(legend=...)
Placing a legend of hypothesis on the plot is a very common thing to
do, yet how to choose the appropriate legend parameter is not that
straightforward.

We provide some commonly used settings for legend, such as
RIGHT/BOTTOM/TOP to the figure as well as the AUTO behavior that we
previously had. The legend parameter also can take a function
for the sake of more flexibility.