Allow custom representative_fn in hypothesis.plot()
So far the representative value for drawing a plot/curve was always
the mean of the individual runs, but we now support and allow
more general representative values such as median or mode. To this end,
a new parameter `representative_fn` has been added.

An accompanying change is that `err_fn` may return a tuple of two
DataFrames that represent the range of the error band explicitly
in the form of (lower, upper) range, in a case a custom representative
value other than the mean is used, as well as a single DataFrame
representing the radius from the mean.