Make SSH tests connect to and read from localhost itself
This way, we don't need to set up and configurean additional remote
SSH server for testing.

The only requirement is a ssh server set up on the localhost, and
SSH key for authentication, so that `ssh localhost` can work.

Github Actions CI should now run the SFTP related tests as well.