Add path_util.session() to cache and reuse SSH/SFTP connections
Handshaking and establishing SSH sessions (for SFTP) is expensive;
consecutive and many numbers of path_util call while loading runs may
take a lot of time if the SSH connection is not cached or re-used.

We add a context manager `path_util.session()`, within which SSH/SFTP
connection resources can be cached and reused for a fast performance.