Make expt's rust extension optional
When rust toolchain is not available, use TensorboardLogReader
as a fallback (which is slow), just for the users or environments
who do not have rust compilers but still want to install expt anyway.

In order to disable rust extensions, an explicit config must be made
during installation by setting `EXPT_DISABLE_RUST=1`.