Make Hypothesis groupby functions work with pandas 2.0+
Since pandas 2.0, DataFrameGroupBy numeric_only defaults to True and
no longer accepts None; a non-numeric data will result in TypeError.

Our sensible default behavior is to ignore all the non-numeric columns
(which will result in a DataFrame column filled with None); to ensure
this behavior with both pandas 1.x and 2.x, we set the default_only=True
as the default parameter for groupby-related methods (mean, max, min,
etc.) of Hypothesis.